+++ packages/core/build.config.ts
@@ -53,7 +53,7 @@ export default defineBuildConfig({
 		'@': resolve(__dirname, './src')
 	},
 	replace: {
-		__PLUGIN_VERSION__: "\"1.4.0\""
+		__PLUGIN_VERSION__: "\"1.4.1\""
 	},
 	rollup: {
 		emitCJS: true,